python_core/requirements.txt
fastapi[all]==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
alembic==1.15.2
asyncpg==0.30.0
pydantic-settings==2.8.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
openai==1.70.0
edge-tts==6.1.3
loguru==0.7.3
psutil==7.0.0
pyautogui==0.9.54
asyncio-mqtt==2.2.0
transformers==4.50.3
torch==2.6.0
numpy==2.2.4
sentence-transformers==3.4.1
pgvector==0.3.6
beautifulsoup4==4.13.3
google-api-python-client==2.160.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
fitbit==0.4.0
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-cov==6.0.0
pytest-mock==3.14.0
